Turn off colors
This is for two reaons:

1.  There have been a number of complaints about the color blue.  For now I
can't get reasonable agreement on which color to make informative messages and
it will differ depending on what the background color is of a person's terminal.

2.  Currently the log is generated simply by splitting stdout and stderr.  That
means that color codes go to the log, which is less than ideal.  We need
separate logging routines that don't send the color codes but for now turning
off color will suffice.

This has been accomplished simply by commenting out the tput assignments at the
beginning of the script.  It is likely that osme form of color support will go
back in at some point so we'll keep the old code around for now.