Force en_US.UTF-8
In order to properly parse output from certain commands we must ensure that
output is in the correct locale and character set.  Previously we had just set
LANG but that doesn't force the issue and some systems may still have issues.
While this change will override a person's foreign language preferences this is
preferable to it causing a failed migration.