Dump DNF cache before collecting system data and dnf update instead of after.
If there's going to be errors due to corrupted repos, etc we want to find out
right away so we can bail before we put the system in an unstable state.  To
this end we dump the dnf cache right away to make sure that dnf is forced to hit
the repos insteada of relying on cached data.