migrate2rocky: fix listing of rpm info files
When the conversion completes, the script prints:

    You may review the following files:

However, there are no files displayed because the find pattern does not
match what `generate_rpm_info()` writes.  It looks for '*-rpms-*' while
the files are named '*-rpm-list-*' instead.

Fix the file pattern used.  While here, replace the use of `find` with a
simpler `printf` to achieve the same result.