migrate2rocky minor improvements (#140)
* migrate2rocky: use $convert_info_dir consistently

We set convert_info_dir=/root/convert.  Use it consistently.

* migrate2rocky: avoid mkdir error with -rV options

Both -r and -V call `generate_rpm_info()` which run `mkdir` to create
the convert directory.  The second call results in a warning:

    mkdir: cannot create directory ‘/root/convert’: File exists

Use `mkdir -p` to avoid the spurious warning when the convert directory
exists.

* migrate2rocky: fix listing of rpm info files

When the conversion completes, the script prints:

    You may review the following files:

However, there are no files displayed because the find pattern does not
match what `generate_rpm_info()` writes.  It looks for '*-rpms-*' while
the files are named '*-rpm-list-*' instead.

Fix the file pattern used.  While here, replace the use of `find` with a
simpler `printf` to achieve the same result.